+++ test/verify.test.js
@@ -1,20 +1,20 @@
 var chai = require('chai');
 var sinon = require('sinon');
 var Strategy = require('../lib/strategy');
-var jwt = require('jsonwebtoken');
-
-
-function buildIdToken() {
-  return jwt.sign({some: 'claim'}, 'this is a secret', {
-    issuer: 'https://server.example.com',
-    subject: '248289761001',
-    audience: 's6BhdRkqt3',
-    expiresIn: '1h'
-  });
-};
+var jws = require('jws');
 
 
 describe('verify function', function() {
+  var clock;
+  
+  beforeEach(function() {
+    clock = sinon.useFakeTimers(1311280970000);
+  });
+  
+  afterEach(function() {
+    clock.restore();
+  });
+  
   
   describe('that authenticates', function() {
     
@@ -46,7 +46,17 @@ describe('verify function', function() {
       sinon.stub(strategy._oauth2, 'getOAuthAccessToken').yieldsAsync(null, 'SlAV32hkKG', '8xLOxBtZp8', {
         token_type: 'Bearer',
         expires_in: 3600,
-        id_token: buildIdToken()
+        id_token: jws.sign({
+          header: { alg: 'HS256' },
+          payload: {
+            iss: 'https://server.example.com',
+            sub: '248289761001',
+            aud: 's6BhdRkqt3',
+            exp: Math.floor((Date.now() + 1000000) / 1000),
+            iat: Math.floor(Date.now() / 1000)
+          },
+          secret: 'keyboard cat',
+        })
       });
       
       sinon.stub(strategy._oauth2, 'get').yieldsAsync(null, JSON.stringify({
@@ -82,7 +92,84 @@ describe('verify function', function() {
         })
         .error(done)
         .authenticate();
-    }); // should accept profile to authenticate request
+    }); // should accept issuer and profile to authenticate request
+    
+    it('should accept issuer, profile, and ID token to authenticate request', function(done) {
+      var strategy = new Strategy({
+        issuer: 'https://server.example.com',
+        authorizationURL: 'https://server.example.com/authorize',
+        tokenURL: 'https://server.example.com/token',
+        userInfoURL: 'https://server.example.com/userinfo',
+        clientID: 's6BhdRkqt3',
+        clientSecret: 'some_secret12345',
+        callbackURL: 'https://client.example.org/cb'
+      },
+      function(issuer, profile, idToken, cb) {
+        expect(issuer).to.equal('https://server.example.com');
+        var _raw = profile._raw; delete profile._raw;
+        var _json = profile._json; delete profile._json;
+        expect(profile).to.deep.equal({
+          id: '248289761001',
+          username: 'j.doe',
+          displayName: 'Jane Doe',
+          name: { familyName: 'Doe', givenName: 'Jane', middleName: undefined },
+          emails: [ { value: 'janedoe@example.com' } ]
+        });
+        expect(idToken).to.equal('eyJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwczovL3NlcnZlci5leGFtcGxlLmNvbSIsInN1YiI6IjI0ODI4OTc2MTAwMSIsImF1ZCI6InM2QmhkUmtxdDMiLCJleHAiOjEzMTEyODE5NzAsImlhdCI6MTMxMTI4MDk3MH0.2Y-uXE7I6Gfon1v4mZVCRKIfZJ_I8BGQoedagok5MNk');
+        
+        return cb(null, { id: '248289761001' });
+      });
+      
+      sinon.stub(strategy._oauth2, 'getOAuthAccessToken').yieldsAsync(null, 'SlAV32hkKG', '8xLOxBtZp8', {
+        token_type: 'Bearer',
+        expires_in: 3600,
+        id_token: jws.sign({
+          header: { alg: 'HS256' },
+          payload: {
+            iss: 'https://server.example.com',
+            sub: '248289761001',
+            aud: 's6BhdRkqt3',
+            exp: Math.floor((Date.now() + 1000000) / 1000),
+            iat: Math.floor(Date.now() / 1000)
+          },
+          secret: 'keyboard cat',
+        })
+      });
+      
+      sinon.stub(strategy._oauth2, 'get').yieldsAsync(null, JSON.stringify({
+        sub: '248289761001',
+        name: 'Jane Doe',
+        given_name: 'Jane',
+        family_name: 'Doe',
+        preferred_username: 'j.doe',
+        email: 'janedoe@example.com',
+        picture: 'http://example.com/janedoe/me.jpg'
+      }));
+      
+      chai.passport.use(strategy)
+        .request(function(req) {
+          req.query = {
+            code: 'SplxlOBeZQQYbYS6WxSbIA',
+            state: 'af0ifjsldkj'
+          };
+          req.session = {};
+          req.session['openidconnect:server.example.com'] = {
+            state: {
+              handle: 'af0ifjsldkj'
+            }
+          };
+        })
+        .success(function(user, info) {
+          expect(user).to.deep.equal({ id: '248289761001' });
+          expect(info).to.deep.equal({});
+          
+          expect(strategy._oauth2.getOAuthAccessToken.calledOnce).to.be.true;
+          expect(strategy._oauth2.get.calledOnce).to.be.true;
+          done();
+        })
+        .error(done)
+        .authenticate();
+    }); // should accept issuer, profile, and ID token to authenticate request
     
     it('should accept profile and tokens and authenticate request', function(done) {
       var strategy = new Strategy({
@@ -115,7 +202,17 @@ describe('verify function', function() {
       sinon.stub(strategy._oauth2, 'getOAuthAccessToken').yieldsAsync(null, 'SlAV32hkKG', '8xLOxBtZp8', {
         token_type: 'Bearer',
         expires_in: 3600,
-        id_token: buildIdToken()
+        id_token: jws.sign({
+          header: { alg: 'HS256' },
+          payload: {
+            iss: 'https://server.example.com',
+            sub: '248289761001',
+            aud: 's6BhdRkqt3',
+            exp: Math.floor((Date.now() + 1000000) / 1000),
+            iat: Math.floor(Date.now() / 1000)
+          },
+          secret: 'keyboard cat',
+        })
       });
       
       sinon.stub(strategy._oauth2, 'get').yieldsAsync(null, JSON.stringify({